_semantic/requirements.txt
autogen
azure-ai-inference
AzureOpenAI
#azure-openai
azure-identity
pandas
requests
beautifulsoup4
playwright
autogenstudio
crawl4ai
python-dotenv
semantic-kernel